sysmem: make it runnable through uOFW installer!
I had to revert the "kprxgen" to the original psp-prxgen and add a second segment containing the relocation offset (program type 0x700000A0) so it is recognized by reboot.bin which only handles relocations whose address is specified in a program header. So now psp-kprxgen must be used to compile at least sysmem.
This trick didn't work for loadcore; I'll investigate later.