Configured travis to build and deploy the executables
The executables are rebuilt whenever the utils/ directory is modified or when the commit title contains [exec-build]. I have also added an option to disable rebuilding the documentation with the [docs-skip] directive. When [docs-skip] is in the commit title the documentation is not updated. This commit also makes travis preserve the history of the gh-pages branch.