Minor assembly changes (should be a no-op) (#138)
The "break" instruction is a bit ambiguous in the MIPS assebmler. It has
a 20 bit immediate, but some assemblers treat it like a 10 bit constant
(with 10 LSB as zero) and other as a full 20 bit immediate. The upstream
MIPS binutils use the former while pspdev's branch uses the later.

Using the two immediate format (hi10, lo10) removes this ambiguity. I
also removed "break 0" as simply "break".